+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added support for setting placeholder UserId and GroupId
+
 ### Fixed
 
 - Fixed generation of values-dev.yaml in hack/kind.sh